requirements.txt
openai~=1.75
openai-agents~=0.0
python-dotenv~=1.1
tqdm~=4.67
pytube~=15.0.0
yt-dlp==2025.3.31